Improve precision of OwnershipPercentage property
The precision of the `OwnershipPercentage` property in the
`StockOwnershipConfiguration` class has been improved from
`decimal(4, 2)` to `decimal` with a precision of 5 and a scale
of 2.

Updated the `20250125220755_ChangeReviewStatusToEnum.cs`
migration file to include `DbContext` and `Migration`
attributes, specifying `FundamentalDbContext` and the
migration name.

Added a new migration file
`20250126045347_ImprovePrecisionOfShareHolderProperty.cs` to
alter the `ownership_percentage` column in the
`stock_ownership` table to the new precision settings.

Updated the `20250126045347_ImprovePrecisionOfShareHolderProperty.Designer.cs`
file to reflect the new migration name and changes to the
`OwnershipPercentage` property.

Updated `FundamentalDbContextModelSnapshot.cs` to ensure the
model snapshot is consistent with the new precision settings.