Refactor index data handling and update related services
The changes were necessary to improve the structure and efficiency of index data
handling. By returning a single `IndexResponse` object instead of a list, we
simplify the data model and make it easier to manage. The introduction of the
`IndexResponseItem` class allows for a more organized representation of index
data, which enhances readability and maintainability.

Updating the `fromDate` initialization in `UpdateIndexDataCommandHandler`
ensures that we capture a broader range of historical data. The new
`WhereProductType` method in `SymbolSpec` provides additional filtering
capabilities, making the codebase more flexible and robust.

Including the `UpdateIndexDataCommand` in the `CommonCodalDataHostedService`
ensures that index data is updated periodically, improving the accuracy and
reliability of the service. These changes collectively enhance the overall
functionality and performance of the system.