Update precision and add admin-provided ownership percentage
In order to support larger and more precise ownership percentages, the precision of
the `OwnershipPercentage` property in the `StockOwnershipConfiguration` class has
been increased from `(5, 2)` to `(18, 4)`. Additionally, a new property
`OwnershipPercentageProvidedByAdmin` has been introduced to allow administrators
to provide their own ownership percentage values. This property also has a
precision of `(18, 4)` and is required.

The `20250203203654_AddUrlToStockOwnershipEntity` migration class has been updated
to include the `DbContext` attribute and a new migration name to ensure proper
tracking of changes. A new migration class
`20250204074038_AddPercentageByAdminColumnToStockOwnership` has been created to
apply these changes to the database schema, including altering the
`ownership_percentage` column and adding the new
`ownership_percentage_provided_by_admin` column with a default value of `0m`.

The `20250204074038_AddPercentageByAdminColumnToStockOwnership.Designer.cs` file
and the `FundamentalDbContextModelSnapshot` have been updated to reflect these
changes, ensuring that the entity model and database schema are in sync.