Update project metadata in Directory.Build.props
The changes in the `Directory.Build.props` file are necessary to reflect the
new development team and product focus. The authorship has been updated to
"PeSahm Development Team" to indicate the new team responsible for the project.
The company information has been removed as it is no longer relevant. The
copyright information has been updated to "Mostafa Esmaeili And Amir Mehrizi"
to accurately represent the new ownership. Finally, the product name has been
changed to "Fundamental Platform" to align with the new branding and direction
of the project.