Add OnlyTagged filter to GetNonOperationIncomesRequest
This change introduces a new boolean property `OnlyTagged` to the
`GetNonOperationIncomesRequest` class, allowing users to filter non-operation
incomes based on whether they are tagged. The `NonOperationIncomesRepository`
query logic has been updated to respect this new property, ensuring that when
`OnlyTagged` is set to `true`, only entries with tags are included in the
results. This enhancement provides more granular control over the data
retrieved, improving the efficiency and relevance of the query results for
users who need to focus on tagged entries.